MYSQL_TYPE_LONG

+++ script/mysql/odb.sql
@@ -4,7 +4,7 @@ CREATE TABLE `GIT_ODB` (
  `repo` BIGINT UNSIGNED NOT NULL,
  `oid`  BINARY(20) NOT NULL,
  `type` TINYINT(1) UNSIGNED NOT NULL,
- `size` BIGINT UNSIGNED NOT NULL,
+ `size` INT UNSIGNED NOT NULL,
  `data` LONGBLOB NOT NULL,
  PRIMARY KEY (`repo`,`oid`),
  KEY `type` (`repo`,`type`),
